added alertbox at deleteNote
